+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 ### Changed
 * Changed ActiveStorage to use VIPS on backend ([#1836](https://gitlab.com/yale-sdmp/reservations/issues/1836)).
+* Changed Rails to not return default error page for bad host requests ([#1837](https://gitlab.com/yale-sdmp/reservations/issues/1837)).
 
 ### Removed
 * Removed references to Travis CI ([#1835](https://gitlab.com/yale-sdmp/reservations/issues/1835)).